move acc_utils to acc
