+++ bluepyopt/ephys/morphologies.py
@@ -28,9 +28,9 @@ import pathlib
 import bisect
 import numpy
 from bluepyopt import _arbor as arbor
+from bluepyopt.ephys.acc.acc_utils import ArbLabel
 from bluepyopt.ephys.base import BaseEPhys
 from bluepyopt.ephys.serializer import DictMixin
-from bluepyopt.ephys.acc_utils import ArbLabel
 
 logger = logging.getLogger(__name__)
